requirements.txt
# FastAPI and dependencies - Python 3.13 compatible
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database - Use newer SQLAlchemy version that supports Python 3.13
sqlalchemy[asyncio]>=2.0.25

# Use psycopg (version 3) - Updated to available versions
# psycopg v3 has native asyncio support and works with Python 3.13
psycopg[binary]>=3.2.2
psycopg[pool]>=3.2.2

# Fallback: Keep psycopg2-binary for sync operations
psycopg2-binary>=2.9.9

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2

# Scheduling - Updated for Python 3.13
apscheduler==3.10.4

# Environment management
python-dotenv==1.0.0

# Web server
gunicorn==21.2.0

# Templates (for FastAPI)
jinja2==3.1.2

# Email functionality (optional but recommended)
yagmail==0.15.293

# Additional packages for better Python 3.13 support
typing-extensions>=4.8.0

# Alternative: Use asyncio-compatible packages
aiopg==1.4.0  # Alternative async PostgreSQL driver
